Added Sound Manager (WIP) + Unrelated SFX/Music for testing
Added WIP sound manager that can adjust master volume. Need to further add to it so that it can adjust SFX and music separately and the volume sliders will remain in the same position as it was left when exiting the options scene. I also added sounds and music unrelated to the game just for testing purposes.